Moved natives into seperate file.
