+++ src/vm.c
@@ -5,44 +5,11 @@
 #include "object.h"
 #include "memory.h"
 #include "leb128.h"
+#include "natives.h"
 #include <stdio.h>
 #include <stdarg.h>
 #include <string.h>
-#include <time.h>
 #include <math.h>
-
-static void defineNative(VM* vm, const char* name, size_t arity, NativeFn function);
-static void runtimeError(VM* vm, const char* format, ...);
-
-static Value clockNative(VM* vm, uint8_t argCount, Value* args, bool* hasError) {
-	return NUMBER_VAL((double)clock() / CLOCKS_PER_SEC);
-}
-
-static Value printNative(VM* vm, uint8_t argCount, Value* args, bool* hasError) {
-	for (size_t i = 0; i < argCount; i++) {
-		printf("%s", valueToString(vm, args[i])->chars);
-		printf(" ");
-	}
-	printf("\n");
-	return NULL_VAL;
-}
-
-static Value toStringNative(VM* vm, uint8_t argCount, Value* args, bool* hasError) {
-	return OBJ_VAL(valueToString(vm, args[0]));
-}
-
-static Value reprNative(VM* vm, uint8_t argCount, Value* args, bool* hasError) {
-	return OBJ_VAL(valueToRepr(vm, args[0]));
-}
-
-static Value sqrtNative(VM* vm, uint8_t argCount, Value* args, bool* hasError) {
-	if (!IS_NUMBER(args[0])) {
-		runtimeError(vm, "Expected number as first argument to sqrt.");
-		*hasError = true;
-		return NULL_VAL;
-	}
-	return NUMBER_VAL(sqrt(AS_NUMBER(args[0])));
-}
 
 static void resetStack(VM* vm) {
 	vm->stackTop = vm->stack;
@@ -75,11 +42,7 @@ void initVM(VM* vm) {
 
 	tableSet(vm, &vm->globals, copyString(vm, "NaN", 3), NUMBER_VAL(nan("0")));
 	tableSet(vm, &vm->globals, copyString(vm, "Infinity", 8), NUMBER_VAL(INFINITY));
-	defineNative(vm, "toString", 1, toStringNative);
-	defineNative(vm, "repr", 1, reprNative);
-	defineNative(vm, "clock", 0, clockNative);
-	defineNative(vm, "sqrt", 1, sqrtNative);
-	defineNative(vm, "print", 1, printNative);
+	defineGlobalNatives(vm);
 }
 
 void freeVM(VM* vm) {
@@ -103,7 +66,7 @@ static Value peek(VM* vm, size_t distance) {
 	return vm->stackTop[-1 - distance];
 }
 
-static void runtimeError(VM* vm, const char* format, ...) {
+void runtimeError(VM* vm, const char* format, ...) {
 	va_list args;
 	va_start(args, format);
 	vfprintf(stderr, format, args);
@@ -159,14 +122,6 @@ static void closeUpvalues(VM* vm, Value* last) {
 		upvalue->location = &upvalue->closed;
 		vm->openUpvalues = upvalue->next;
 	}
-}
-
-static void defineNative(VM* vm, const char* name, size_t arity, NativeFn function) {
-	push(vm, OBJ_VAL(copyString(vm, name, strlen(name))));
-	push(vm, OBJ_VAL(newNative(vm, arity, function)));
-	tableSet(vm, &vm->globals, AS_STRING(vm->stack[0]), vm->stack[1]);
-	pop(vm);
-	pop(vm);
 }
 
 static bool call(VM* vm, ObjClosure* closure, uint8_t argCount) {
