+++ src/vm.h
@@ -42,5 +42,6 @@ typedef enum {
 void initVM(VM* vm);
 void freeVM(VM* vm);
 InterpreterResult interpret(VM* vm, const char* source);
+void runtimeError(VM* vm, const char* format, ...);
 void push(VM* vm, Value value);
 Value pop(VM* vm);
